Add files via upload
As players progress through the game, they will encounter portals that force them to wait for a while before they can progress or make an in-app purchase. In this project, we will analyze the result of A/B testing where the first portal in Cookie Cats has been moved from level 30 to level 40. In particular, we will analyze the impact on player retention.